server/requirements.txt
alembic==1.14.1
blinker==1.8.2
certifi==2025.4.26
click==8.1.8
distlib==0.3.9
filelock==3.16.1
Flask==3.0.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==2.1.5
packaging==25.0
pipenv==2024.4.1
platformdirs==4.3.6
pyotp==2.9.0
python-dotenv==1.0.1
SQLAlchemy==2.0.41
typing_extensions==4.13.2
virtualenv==20.30.0
Werkzeug==3.0.6
zipp==3.20.2
cloudinary
